refactor: 응답용 객체의 패키지 구조 변경
